add new handle_key_error
KeyError now is manually handled as it is quite a simple error with only one possible origin (user tried to access a key of a dict that does not exist).
The new handle has a default message that will be filled with the apporpriate information collected from the offending line and the error message.