Add cache to functions that calls StackOverflow API (#33)
* Add cache to funcions that calls stackoverflow API

This feature is provided by the filecache package. cache is stored on the same level ofthe file that need it. Cache has a due date of 7 days.

* add cache files to .gitignore

To avoid someone accidentaly commiting cache files while developing pycee

* add filecache to requirements.txt

* Update tests

To cache function calls the return value must be hashable. So tests now reflect this change