Add tests for syntax error handlers
This covers both methods that handles syntax errors. The strategy is that if the user has a generic syntax error, give him/her some hints on fixing simples syntax errors. If the user is facing a specific syntax error (which can be checked using the traceback error message) then let's search stackoverflow for an possible answer