Setup test coverage report (#46)
we need to install pytest-cov to generate coverage information, Github actions can do this for us running pip install pytest-cov, When this happens, pytest-cov generates some files and folders (.coverage, coverage.xml). To interpret the data in these files we will use codecov, which will read them and show us some coverage information. The most important thing here is the percentage of coverage, we don't need much more data about it.
a badge will be available in the readme.